🐛Source-mssql: aligned regular and cdc syncs and its datatype tests (#14379)
* [13606] source-mssql: Merged datatype tests for normal and CDC syncs
