+++ airbyte-config/init/src/main/resources/seed/source_specs.yaml
@@ -4971,7 +4971,7 @@
     supportsNormalization: false
     supportsDBT: false
     supported_destination_sync_modes: []
-- dockerImage: "airbyte/source-mssql:0.4.8"
+- dockerImage: "airbyte/source-mssql:0.4.9"
   spec:
     documentationUrl: "https://docs.airbyte.io/integrations/destinations/mssql"
     connectionSpecification:
