+++ airbyte-db/db-lib/src/main/java/io/airbyte/db/DataTypeUtils.java
@@ -12,6 +12,7 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.function.Function;
 import org.slf4j.Logger;
@@ -29,6 +30,7 @@ public class DataTypeUtils {
   public static final DateTimeFormatter TIMESTAMP_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSSSS");
   public static final DateTimeFormatter TIMETZ_FORMATTER = DateTimeFormatter.ofPattern("HH:mm:ss.SSSSSSXXX");
   public static final DateTimeFormatter TIMESTAMPTZ_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSSSSXXX");
+  public static final DateTimeFormatter OFFSETDATETIME_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSSSSSS XXX");
 
   // wrap SimpleDateFormat in a function because SimpleDateFormat is not threadsafe as a static final.
   public static DateFormat getDateFormat() {
@@ -98,6 +100,10 @@ public class DataTypeUtils {
 
   public static String toISO8601String(final LocalDateTime date) {
     return date.format(DateTimeFormatter.ofPattern(DATE_FORMAT_PATTERN));
+  }
+
+  public static String toISO8601String(final OffsetDateTime date) {
+    return date.format(OFFSETDATETIME_FORMATTER);
   }
 
   public static String toISO8601String(final Duration duration) {
