+++ airbyte-integrations/connectors/source-mssql-strict-encrypt/Dockerfile
@@ -16,5 +16,5 @@ ENV APPLICATION source-mssql-strict-encrypt
 
 COPY --from=build /airbyte /airbyte
 
-LABEL io.airbyte.version=0.4.5
+LABEL io.airbyte.version=0.4.6
 LABEL io.airbyte.name=airbyte/source-mssql-strict-encrypt
