+++ airbyte-integrations/connectors/source-mssql/Dockerfile
@@ -16,5 +16,5 @@ ENV APPLICATION source-mssql
 
 COPY --from=build /airbyte /airbyte
 
-LABEL io.airbyte.version=0.4.8
+LABEL io.airbyte.version=0.4.9
 LABEL io.airbyte.name=airbyte/source-mssql
